+++ doc/changelog.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v5.2.2-alpha] - (7/25/2023)
+
+### Added
+- VEX API Endpoint
+- VEX API test `GenerateVEXAPITest`
+- `VEXResult.java`
+  - Class that holds both the VEX Object and any errors that occurred for the API endpoint
+- Test SBOMs that contain vulnerable components
+
+### Changed
+
+
 ## [v5.2.1-alpha] - (7/21/2023)
 
 ### Added
@@ -15,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `comparison/utils/Utils.java`
 
 ### Changed
-- `Merger.java`abstract class and overall architecture
+- Updated `API.md` documentation with the VEX endpoint
 
 ## [v5.2.0-alpha] - (7/20/2023)